# Conflicts:
#	.idea/misc.xml
#	src/Controller/MainViewController.java

+++ src/Controller/AddPatientController.java
@@ -33,7 +33,9 @@ public class AddPatientController {
         this.model = model;
     }
 
-
+    /*
+     * Called when the user clicks "Cancel".
+     */
     @FXML
     private void handleCancel() {
         /* Close the window */
@@ -45,7 +47,7 @@ public class AddPatientController {
     @FXML
     private void handleSubmit() {
 
-        Patient patient = new Patient(firstname.getText(), lastname.getText(), age.getText(), illness.getText());
+        Patient patient = new Patient(firstname.getText(), lastname.getText(),Integer.valueOf(age.getText()).intValue(), illness.getText());
 
         model.addPatient(patient);
